changing datasource to plugin
